# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: AideaCreative Studio Foto API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /produk:
    get:
      operationId: listProduk
      tags: [produk]
      summary: Daftar produk toko
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Produk"
    post:
      operationId: createProduk
      tags: [produk]
      summary: Buat produk baru
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProdukBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Produk"
  /produk/{id}:
    get:
      operationId: getProduk
      tags: [produk]
      summary: Detail produk
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Produk"
    put:
      operationId: updateProduk
      tags: [produk]
      summary: Update produk
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProdukBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Produk"
    delete:
      operationId: deleteProduk
      tags: [produk]
      summary: Hapus produk
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /promo:
    get:
      operationId: listPromo
      tags: [promo]
      summary: Daftar promo
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Promo"
    post:
      operationId: createPromo
      tags: [promo]
      summary: Buat promo baru
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePromoBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Promo"
  /promo/{id}:
    put:
      operationId: updatePromo
      tags: [promo]
      summary: Update promo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePromoBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Promo"
    delete:
      operationId: deletePromo
      tags: [promo]
      summary: Hapus promo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /testimoni:
    get:
      operationId: listTestimoni
      tags: [testimoni]
      summary: Daftar testimoni yang disetujui
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Testimoni"
    post:
      operationId: createTestimoni
      tags: [testimoni]
      summary: Tambah testimoni baru
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTestimoniBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Testimoni"
  /testimoni/{id}:
    delete:
      operationId: deleteTestimoni
      tags: [testimoni]
      summary: Hapus testimoni
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Produk:
      type: object
      properties:
        id:
          type: string
        namaProduk:
          type: string
        deskripsi:
          type: string
        harga:
          type: integer
        stok:
          type: integer
        kategori:
          type: string
          enum: [cetak_foto, frame, album, photobook, merchandise]
        ukuran:
          type: string
        gambarUrl:
          type: array
          items:
            type: string
        isAktif:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - namaProduk
        - deskripsi
        - harga
        - stok
        - kategori
        - gambarUrl
        - isAktif
        - createdAt
    CreateProdukBody:
      type: object
      properties:
        namaProduk:
          type: string
        deskripsi:
          type: string
        harga:
          type: integer
        stok:
          type: integer
        kategori:
          type: string
          enum: [cetak_foto, frame, album, photobook, merchandise]
        ukuran:
          type: string
        gambarUrl:
          type: array
          items:
            type: string
        isAktif:
          type: boolean
      required:
        - namaProduk
        - deskripsi
        - harga
        - stok
        - kategori
        - isAktif
...
    Testimoni:
      type: object
      properties:
        id:
          type: string
        pelangganId:
          type: string
        bookingId:
          type: string
        rating:
          type: integer
        komentar:
          type: string
        namaTampil:
          type: string
        fotoUrl:
          type: string
        isApproved:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - rating
        - komentar
        - namaTampil
        - isApproved
        - createdAt
    CreateTestimoniBody:
      type: object
      properties:
        namaTampil:
          type: string
        rating:
          type: integer
        komentar:
          type: string
        fotoUrl:
          type: string
        bookingId:
          type: string
      required:
        - namaTampil
        - rating
        - komentar
...
    Promo:
      type: object
      properties:
        id:
          type: string
        judul:
          type: string
        deskripsi:
          type: string
        badge:
          type: string
          nullable: true
        gambarUrl:
          type: string
          nullable: true
        link:
          type: string
          nullable: true
        ctaLabel:
          type: string
          nullable: true
        warna:
          type: string
          nullable: true
        tampilMarquee:
          type: boolean
        tampilCard:
          type: boolean
        tanggalMulai:
          type: string
          nullable: true
        tanggalBerakhir:
          type: string
          nullable: true
        isAktif:
          type: boolean
        urutan:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - judul
        - deskripsi
        - tampilMarquee
        - tampilCard
        - isAktif
        - urutan
        - createdAt
    CreatePromoBody:
      type: object
      properties:
        judul:
          type: string
        deskripsi:
          type: string
        badge:
          type: string
          nullable: true
        gambarUrl:
          type: string
          nullable: true
        link:
          type: string
          nullable: true
        ctaLabel:
          type: string
          nullable: true
        warna:
          type: string
        tampilMarquee:
          type: boolean
        tampilCard:
          type: boolean
        tanggalMulai:
          type: string
          nullable: true
        tanggalBerakhir:
          type: string
          nullable: true
        isAktif:
          type: boolean
        urutan:
          type: integer
      required:
        - judul
        - deskripsi
        - isAktif